Huge overhaul of old code
Still a far way from being perfect, but a lot of stuff updated:
* Favicon added
* Settings (most of them anyway) are stored in the new ImmutableSettings
* Hopefully all UI stuff is running on the correct thread now
* Improved UPnP listing (only list devices that can _PROBABLY_ do port mapping)
* A lot of other stuff
* Probably introduced some new bugs too, yay!

+++ src/main/java/se/bes/mhfs/cli/CLIStart.java
@@ -30,7 +30,7 @@ public class CLIStart {
     public static void main(String[] args) {
         if(args.length == 2){
             int port = Integer.parseInt(args[0]);
-            double speed = Integer.parseInt(args[1]);
+            int speed = Integer.parseInt(args[1]);
             new CLIRunner(port, speed);
         }else{
             new CLIRunner();
